third_party/WebKit/LayoutTests/external/wpt/mixed-content/allowed/meta-csp/same-host-https/audio-tag/top-level/no-redirect/allowed.https-expected.txt

This is a testharness.js-based test.
FAIL opt_in_method: meta-csp\n                                 origin: same-host-https\n                                 source_scheme: https\n                                 context_nesting: top-level\n                                 redirection: no-redirect\n                                 subresource: audio-tag\n                                 expectation: allowed assert_equals: The triggered event should match 'allowed'. expected "allowed" but got "blocked"
Harness: the test ran to completion.
